feat: Add GitHub MCP integration and refactor test structure - Added GitHub MCP server configuration with Docker - Configured GitHub Personal Access Token for API access - Refactored Infrastructure integration tests structure - Moved test files to Services directory for better organization - Updated GlobalUsings for test projects - Ready for AI-powered GitHub repository management
