+++ tests/ECommerce.Infrastructure.IntegrationTests/GlobalUsings.cs
@@ -1,13 +1,9 @@
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Caching.Distributed;
-using Microsoft.Extensions.Caching.Memory;
-using Microsoft.Extensions.Options;
-using Moq;
-using Xunit;
-using FluentAssertions;
-using ECommerce.Infrastructure.Services;
-using ECommerce.Application.Interfaces;
-using ECommerce.Application.Services;
-using ECommerce.Domain.Entities;
+global using Microsoft.Extensions.Caching.Distributed;
+global using Microsoft.Extensions.Caching.Memory;
+global using Microsoft.Extensions.Options;
+global using Moq;
+global using Xunit;
+global using FluentAssertions;
+global using ECommerce.Infrastructure.Services;
+global using ECommerce.Application.Interfaces;
+global using ECommerce.Domain.Entities;
